fastapi/requirements.txt
fastapi
uvicorn
websockets
openai
litellm
pyyaml
firecrawl-py